cleanup old tests
